+++ pkg/registry/apis/provisioning/jobs/export/exporter.go
@@ -1,4 +1,4 @@
-package jobs
+package export
 
 import (
 	"context"
